Move get_client_ip from api.views to api.utils (new)
