+++ backend/api/utils.py
@@ -1,0 +1,10 @@
+from rest_framework.request import Request
+
+
+def get_client_ip(request: Request):
+    x_forwarded_for = request.META.get('HTTP_X_FORWARDED_FOR')
+    if x_forwarded_for:
+        ip = x_forwarded_for.split(',')[0]
+    else:
+        ip = request.META.get('REMOTE_ADDR')
+    return ip
